requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
Django==2.2
djangorestframework==3.9.4
idna==2.8
importlib-metadata==0.17
importlib-resources==1.0.2
lxml==4.3.3
more-itertools==7.0.0
pluggy==0.12.0
py==1.8.0
pytest==4.5.0
pytest-cov==2.7.1
pytest-django==3.4.8
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.25.2
wcwidth==0.1.7
zipp==0.5.1
